access bundled tex templates properly

+++ pycvc/samples.py
@@ -367,18 +367,22 @@ class _WQSample_Mixin(wqio.core.samples._basic_wq_sample):
         texpath = self._make_ISR_document(version=version)
 
         texdir, texdoc = os.path.split(texpath)
-        with LaTeXDirecory(texdir) as TEXDIR:
-            # use ``pdflatex to compile the document
-            tex = subprocess.call(['pdflatex', texdoc, '--quiet'], shell=False,
-                                  stdout=subprocess.PIPE,
-                                  stderr=subprocess.PIPE)
 
-        if clean:
-            extensions = ['aux', 'log', 'nav', 'out', 'snm', 'toc']
-            for ext in extensions:
-                junkfiles = glob.glob('*.{}'.format(ext))
-                for junk in junkfiles:
-                    os.remove(junk)
+        if wqio.testing.checkdep_tex() is not None:
+            with LaTeXDirecory(texdir) as TEXDIR:
+                # use ``pdflatex to compile the document
+                tex = subprocess.call(['pdflatex', texdoc, '--quiet'], shell=False,
+                                      stdout=subprocess.PIPE,
+                                      stderr=subprocess.PIPE)
+
+            if clean:
+                extensions = ['aux', 'log', 'nav', 'out', 'snm', 'toc']
+                for ext in extensions:
+                    junkfiles = glob.glob('*.{}'.format(ext))
+                    for junk in junkfiles:
+                        os.remove(junk)
+        else:
+            tex = None
 
         return tex
